+++ jacktrade/__init__.py
@@ -1,0 +1,8 @@
+# Because each submodule is small, there is no need to further subdivide imports
+# into categories. Therefore, make everything available from the top-level module,
+# like a Swiss army knife.
+from .benchmark import *
+from .buffers import *
+from .collections import *
+from .multicore import *
+from .pickler import *