+++ HISTORY.md
@@ -1,5 +1,10 @@
 # Release History
 
+## 0.2.0 (2023-08-08)
+### Improvements
+- Added `StringBuffers` class.
+- All utilities are available for import from the top-level module.
+
 ## 0.1.2 (2023-08-01)
 ### Improvements
 - Chunk size is infinite if set to None.